backend/requirements.txt
flask
flask-cors
pdfplumber
python-docx
google-generativeai
